Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

BBModels.BBModel

BBModels.BBModelMeta

BBModels.Problem

BBModels.Problem

BBModels.ProblemMetrics

BBModels.cost

BBModels.get_counters

BBModels.get_cpu_frequency

BBModels.get_id

BBModels.get_memory

BBModels.get_nlp

BBModels.get_pb_id

BBModels.get_status

BBModels.get_times

BBModels.is_failure

BBModels.memory_only

BBModels.normalize_times!

BBModels.obj_cat

BBModels.obj_nomad

BBModels.random_search

BBModels.sumfc

BBModels.time_only

BBModels.types_indices

NLPModels.obj

​

BBModels.BBModel — Type

Mutable struct BBModel

Represents a black box optimization problem that follows the NLPModel API.

The following constructors are available:

BBModel(parameter_set, problems, solver_function, f; kwargs...)
BBModel(parameter_set, problems, solver_function, f, c, lcon, ucon; kwargs...)

parameter_set::AbstractParameterSet: structure containing parameters information;

problems::Vector{AbstractNLPModel}: set of problem to run the benchmark on;

solver_function::Function: function that takes an AbstractNLPModel and a AbstractParameterSet and returns a GenericExecutionStats.

f::Function: Given a Vector{ProblemMetrics} returns a score as a Float64 (examples are time_only, memory_only, sumfc);

For constrained problems:

lcon ≤ c(x) ≤ ucon

c::Function: constraint function;

lcon::AbstractVector: lower bound on the constraint;

ucon::AbstractVector: upper bound on the constraint.

Additional keyword arguments are:

subset::NTuple{N, Symbol}: subset of parameters to be considered (by default all parameters from parameter_set);

x0::AbstractVector: initial values for the parameters (by default Float64.(values(parameter_set)));

lvar::AbstractVector: lower bound on the the parameters (by default Float64.(lower_bounds(parameter_set)));

uvar::AbstractVector: upper bound on the the parameters (by default Float64.(lower_bounds(parameter_set)));

name::String: name of the problem (by default: "Generic").

Note that if x0 is not provided, the computations are run in Vector{Float64}.

BBModels.BBModelMeta — Type

BBModelMeta

A composite type that represents the main features of the blackbox optimization problem of parameters.

BBModelMeta(parameter_set::P [, subset::NTuple{N, Symbol} = fieldnames(P)])

Create an BBModelMeta of all parameters or a subset on a parameter_set of type P <: AbstractParameterSet.

BBModelMeta contains the following attributes:

nvar: number of variables;

x_n::Vector{String}: names of the parameters;

icat::Vector{Int}: indices of categorical variables;

ibool::Vector{Int}: indices of boolean variables;

iint::Vector{Int}: indices of integer variables;

ifloat::Vector{Int}: indices of real variables.

BBModels.Problem — Type

Mutable struct Problem

Keep track of an instance of an AbstractNLPModel in a distributed context by giving each instance an id and a non-negative weight.

The following constructor is available:

Problem(id::Int, nlp::Union{AbstractNLPModel, Function}, weight::Float64)

Assign an id and a weight to the problem nlp. Note that nlp can be given as an Function returning an AbstractNLPModel once evaluated.

BBModels.Problem — Method

Constructor of a Problem. Takes an id and an AbstractNLPModel.

BBModels.ProblemMetrics — Type

ProblemMetrics

Structure that contains metrics of a given solver applied to an AbstractNLPModel.

The following metrics are stored:

pb_id::Int: identifier of the problem solved, see Problem;

times::Vector{Float64}: Execution time of the attempts of solving the nlp;

memory::Int: Memory allocated to solve the nlp;

status::Symbol: Status of the nlp after solve, see SolverCore.jl's documentation;

counters::Counters: Counters of the nlp, see NLPModels.jl's documentation.

BBModels.cost — Method

cost(nlp::BBModel, p::Problem; seconds = 10.0, samples = 1, evals = 1)

For a given problem p::Problem, it returns a ProblemMetrics containing the benchmark's results of nlp.solver_function. The keyword arguments are parameters for the benchmark.

BBModels.get_counters — Method

Returns the Counters related the problem linked to this ProblemMetrics instance.

BBModels.get_cpu_frequency — Method

Returns the CPU frequency in GHz of the host executing this function.

BBModels.get_id — Method

Returns the id of a Problem.

BBModels.get_memory — Method

Returns the memory allocated in Mb to solve the problem linked to this ProblemMetrics instance.

BBModels.get_nlp — Method

Returns the AbstractNLPModel of a Problem.

BBModels.get_pb_id — Method

Returns the id of the problem linked to this ProblemMetrics instance.

BBModels.get_status — Method

Returns the problem linked to this ProblemMetrics instance is solved.

BBModels.get_times — Method

Returns the times required in seconds to solve the problem linked to this ProblemMetrics instance.

BBModels.is_failure — Method

is_failure(status::Symbol)
is_failure(stats::AbstractExecutionStats)

Returns true if a status is considered a failure. SolverCore.show_statuses() return the list of all possible statuses.

BBModels.memory_only — Method

memory_only(vec_metric::Vector{ProblemMetrics}; penalty::Float64 = 5.0)

Return the memory used in p_metric. Unsolved problems are penalyzed by a penalty factor.

BBModels.normalize_times! — Method

Modifies in place a Vector of times by dividing each value by the frequency of the CPU in Mz.

BBModels.obj_cat — Method

obj_cat(nlp::BBModel, x::AbstractVector; kwargs...)

Objective function of the BBModel. The difference with obj is that x contains all the parameters (including categorical variables). Therefore, x is of length nlp.bb_meta.nvar.

The keyword arguments are passed to cost.

BBModels.obj_nomad — Method

obj_nomad(nlp::BBModel, x::AbstractVector, problems::Vector{Problem}; kwargs...)

Objective function of the BBModel to be used with SolverTuning.jl. The difference with obj is the positional parameter problems. Therefore, x is of length nlp.meta.nvar.

The keyword arguments are passed to cost.

BBModels.random_search — Method

random_search(model::BBModel; verbose = 0, max_time = 30.0, max_iter = 10)

Generate random values of the parameter set.

Keywords arguments:

check_cache :: Bool = true: Check if x has already been evaluated;

verbose :: Integer = 0: Print iteration information if >0;

max_time :: Float64 = 30.0: time limit in seconds;

max_iter :: Integer = 10: maximum number of iterations.

BBModels.sumfc — Method

sumfc(vec_metric::Vector{ProblemMetrics}; penalty::Float64 = 5.0)

Return the sum of the evaluations of the objective function and constraint function. Unsolved problems are penalyzed by a penalty factor.

BBModels.time_only — Method

time_only(vec_metric::Vector{ProblemMetrics}; penalty::Float64 = 5.0)

Return the median time, if more than one solve, of p_metric. Unsolved problems are penalyzed by a penalty factor.

BBModels.types_indices — Method

types_indices(parameter_set::P [, subset::NTuple{N, Symbol} = fieldnames(P)])

Return the set of indices of categorical, boolean, integer and real parameters within the subset of fields in P <: AbstractParameterSet.

NLPModels.obj — Method

obj(nlp::BBModel, x::AbstractVector; kwargs...)

Objective function of the BBModel. The difference with obj_cat is that x contains only the numerical parameters (excluding categorical variables). Therefore, x is of length nlp.meta.nvar.

The keyword arguments are passed to cost.